requirements.txt
python-dotenv
Appium-Python-Client